+++ Sources/BroadMonetization/Domain/Checkout/DecodedRUCatalogProduct.swift
@@ -2,8 +2,10 @@ struct DecodedRUCatalogProduct: Decodable {
     let catalogProductID: RUCatalogProductID
     let kind: RUCatalogProductKind
     let appStoreProductID: ProductID?
+    let title: String?
     let price: Money?
     let displayPrice: String?
     let subscriptionPeriod: SubscriptionPeriod
     let supportedMethods: [CheckoutMethod]
+    let credits: Int?
 }